# src/resources/swagger-def.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Studies Service",
    "version": "1.0.0"
  },
  "security": [
    {
      "bearerAuth": [
        "STUDY-${programId}.WRITE"
      ]
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "tags": [
          "Health"
        ],
        "summary": "get status of app",
        "responses": {
          "200": {
            "description": "app is healthy",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "500": {
            "description": "app is not healthy"
          }
        }
      }
    },
    "/studies": {
      "get": {
        "tags": [
          "Studies"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "songId",
            "schema": {
              "type": "string"
            }
          }
        ],
        "summary": "Get all valid studies from song and their users from ego",
        "responses": {
          "200": {
            "description": "Boolean value of whether submission system is disabled or enabled"
          }
        }
      },
      "post": {
        "tags": [
          "Studies"
        ],
        "summary": "Create new study in song and related entities in ego",
        "description": "Create study in song, study group in ego and study policy in ego",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStudyReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The new study.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateStudyRes"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/studies/submitters": {
      "post": {
        "tags": [
          "Studies"
        ],
        "summary": "Add submitters to an existing study's group",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddSubmittersReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The new study.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddSubmittersRes"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "delete": {
        "tags": [
          "Studies"
        ],
        "summary": "Remove a submitter from a study group",
        "parameters": [
          {
            "in": "query",
            "name": "studyId",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "submitter",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The new study.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RemoveSubmitterRes"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "tags": [
    {
      "name": "Health",
      "description": "Service status monitoring"
    },
    {
      "name": "Studies",
      "description": "Studies management endpoints"
    }
  ],
  "components": {
    "responses": {
      "GenericSuccess": {
        "description": "Endpoint finished successfully"
      },
      "ServiceUnavailableError": {
        "description": "Endpoint service unavailable"
      },
      "UnauthorizedError": {
        "description": "Access token is missing or invalid"
      },
      "ServerError": {
        "description": "Server error"
      },
      "ForbiddenError": {
        "description": "Access token has incorrect scopes"
      },
      "NotFoundError": {
        "description": "Requested resource was not found"
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "CreateStudyReq": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "organization": {
            "type": "string"
          },
          "songId": {
            "type": "string"
          },
          "studyId": {
            "type": "string"
          }
        }
      },
      "AddSubmittersReq": {
        "type": "object",
        "properties": {
          "studyId": {
            "type": "string"
          },
          "submitters": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "Studies": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/Study"
        }
      },
      "Study": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "studyId": {
            "type": "string"
          },
          "organization": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "submitters": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "songId": {
            "type": "string"
          }
        }
      },
      "CreateStudyRes": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/Study"
          }
        }
      },
      "AddSubmittersRes": {
        "type": "object",
        "properties": {
          "studyId": {
            "type": "string"
          },
          "submitters": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "RemoveSubmitterRes": {
        "type": "object",
        "properties": {
          "studyId": {
            "type": "string"
          },
          "submitter": {
            "type": "string"
          }
        }
      },
      "ServiceError": {
        "type": "object",
        "required": [
          "success",
          "error"
        ],
        "properties": {
          "success": {
            "type": "boolean"
          },
          "error": {
            "type": "object",
            "required": [
              "type"
            ],
            "properties": {
              "message": {
                "type": "string"
              },
              "type": {
                "type": "string",
                "enum": [
                  "STUDY_NOT_FOUND",
                  "SUBMITTERS_NOT_FOUND",
                  "STUDY_ALREADY_EXISTS",
                  "SUBMITTERS_ALREADY_IN_STUDY",
                  "SUBMITTER_NOT_IN_STUDY",
                  "FAILED_TO_CREATE_STUDY",
                  "FAILED_TO_REMOVE_SUBMITTER_FROM_STUDY",
                  "FAILED_TO_ADD_SUBMITTERS_TO_STUDY",
                  "UNAUTHORIZED",
                  "FORBIDDEN",
                  "UNKNOWN"
                ]
              }
            }
          }
        }
      }
    }
  }
}